+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Add field type textarea
 - Update UI configuration for large feature title
 - Fix #1 #2
+- add static method getFirstValues in model FeatureType